Adds better TypeScript definitions for Angular  (#155)
