+++ packages/okta-angular/src/okta/services/okta.service.ts
@@ -10,12 +10,12 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import { Inject, Injectable, Optional } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { Router, NavigationExtras } from '@angular/router';
 
-import { OKTA_CONFIG } from './okta.config';
+import { OKTA_CONFIG, OktaConfig } from '../models/okta.config';
 
-import packageInfo from './packageInfo';
+import packageInfo from '../packageInfo';
 
 /**
  * Import the okta-auth-js library
@@ -24,11 +24,11 @@ import * as OktaAuth from '@okta/okta-auth-js';
 
 @Injectable()
 export class OktaAuthService {
-    private oktaAuth;
-    private config;
+    private oktaAuth: OktaAuth;
+    private config: OktaConfig;
 
-    constructor(@Inject(OKTA_CONFIG) private auth, private router: Router) {
-      const missing: any[] = [];
+    constructor(@Inject(OKTA_CONFIG) private auth: OktaConfig, private router: Router) {
+      const missing: string[] = [];
 
       if (!auth.issuer) {
         missing.push('issuer');
@@ -67,14 +67,14 @@ export class OktaAuthService {
     /**
      * Returns the OktaAuth object to handle flows outside of this lib.
      */
-    getOktaAuth() {
+    getOktaAuth(): OktaAuth {
       return this.oktaAuth;
     }
 
     /**
      * Checks if there is a current accessToken in the TokenManager.
      */
-    isAuthenticated() {
+    isAuthenticated(): boolean {
       return !!this.oktaAuth.tokenManager.get('accessToken');
     }
 
@@ -98,7 +98,7 @@ export class OktaAuthService {
      * Returns user claims from the /userinfo endpoint if an
      * accessToken is provided or parses the available idToken.
      */
-    async getUser() {
+    async getUser(): Promise<object> {
       const accessToken = this.oktaAuth.tokenManager.get('accessToken');
       const idToken = this.oktaAuth.tokenManager.get('idToken');
       if (accessToken && idToken) {
@@ -115,7 +115,7 @@ export class OktaAuthService {
     /**
      * Returns the configuration object used.
      */
-    getOktaConfig(){
+    getOktaConfig(): OktaConfig {
       return this.config;
     }
 
@@ -156,7 +156,7 @@ export class OktaAuthService {
     /**
      * Parses the tokens from the callback URL.
      */
-    async handleAuthentication() {
+    async handleAuthentication(): Promise<void> {
       const tokens = await this.oktaAuth.token.parseFromUrl();
       tokens.forEach(token => {
         if (token.idToken) {
@@ -181,7 +181,7 @@ export class OktaAuthService {
      * Clears the user session in Okta and removes
      * tokens stored in the tokenManager.
      */
-    async logout() {
+    async logout(): Promise<void> {
       this.oktaAuth.tokenManager.clear();
       await this.oktaAuth.signOut();
     }
@@ -190,7 +190,7 @@ export class OktaAuthService {
      * Scrub scopes to ensure 'openid' is included
      * @param scopes
      */
-    scrubScopes(scopes: string) {
+    scrubScopes(scopes: string): string {
       if (!scopes) {
         return 'openid email';
       }
